backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-docx==1.2.0
python-multipart==0.0.20
sqlalchemy==2.0.43
jinja2==3.1.4
aiofiles==24.1.0
